[refactor] 로또 가게 객체 구조 수정
LottoStore 객체는 static 메서드만을 가지며
LottoStore 객체를 생성하지 않아도 작동할 수 있도록 변경

LottoStore의 메서드는 BuyingStrategy 인터페이스를 인자로 받아
여러가지 방법으로 표를 생성할 수 있도록 변경
