+++ src/main/java/domain/buyinginformation/Money.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.buyinginformation;
 
 public class Money {
